Extract class refactor Airbrake::Notice::JsonBuilder
